--- src/utility/ble_spi.h
@@ -1,13 +1,0 @@
-#pragma once
-
-enum {
-
-
-	BLE_BEGIN 				= 0x4A,
-	BLE_END 				= 0x4B,
-	BLE_AVAILABLE			= 0x4C,
-	BLE_PEEK 				= 0x4D,
-	BLE_READ 				= 0x4E,
-	BLE_WRITE 				= 0x4F,
-
-};
